ACS-3015 moved MockUserNotifier to test folder
